+++ crates/eventwork/examples/server.rs
@@ -78,7 +78,7 @@ struct Player(ConnectionId);
 fn handle_connection_events(
     mut commands: Commands,
     net: Res<Network<TcpProvider>>,
-    mut network_events: EventReader<NetworkEvent>,
+    mut network_events: MessageReader<NetworkEvent>,
 ) {
     for event in network_events.read() {
         if let NetworkEvent::Connected(conn_id) = event {
@@ -96,7 +96,7 @@ fn handle_connection_events(
 
 // Receiving a new message is as simple as listening for events of `NetworkData<T>`
 fn handle_messages(
-    mut new_messages: EventReader<NetworkData<shared::UserChatMessage>>,
+    mut new_messages: MessageReader<NetworkData<shared::UserChatMessage>>,
     net: Res<Network<TcpProvider>>,
 ) {
     for message in new_messages.read() {
@@ -111,9 +111,9 @@ fn handle_messages(
     }
 }
 
-// Handle messages sent via OutboundMessage EventWriter
+// Handle messages sent via OutboundMessage MessageWriter
 fn handle_outbound_messages(
-    mut new_messages: EventReader<NetworkData<shared::OutboundTestMessage>>,
+    mut new_messages: MessageReader<NetworkData<shared::OutboundTestMessage>>,
 ) {
     for message in new_messages.read() {
         let user = message.source();